fixin and implementing Framework compiancy on Animation Graph :
Étape 1 intégrée sur le checkout principal.
**Ce qui a été aligné**
- `AnimationGraphRenderer` : état canvas sérialisable/restaurable ajouté
- `AnimationGraphDocument` : `GetType()` passe à `ANIMATION_GRAPH`
- `IGraphDocument` : ajout de `ANIMATION_GRAPH`
- `CanvasToolbarRenderer` : filtre `*.ani.json`, dossier `./Gamedata/Animation/`, `SaveFileType::AnimationGraph`

Étape 2 corrigée et appliquée.
**Fait**
- `AnimationGraphFrameworkDocument::GetType()` passe à `ANIMATION_GRAPH`
- `SetFilePath()` n’est plus un no-op
- `TabManager` journalise l’affectation du document AnimationGraph
- todo `anim-document-adapter` marqué **done**
**Note**
- J’ai gardé le câblage minimal, sans refactorer le flux métier.

+++ Source/BlueprintEditor/EntityPrefabEditor/PrefabCanvas.h
@@ -944,7 +944,7 @@ namespace Olympe
                     ImVec2 portScreen = m_canvasEditor->CanvasToScreen(ImVec2(portCanvasPos.x, portCanvasPos.y));
 
                     // Pulse animation based on time
-                    float t = ImGui::GetTime();
+                    float t = static_cast<float>(ImGui::GetTime());
                     float pulse = 0.5f + 0.5f * sinf(t * 6.0f);
                     float radius = 8.0f + pulse * 4.0f;
                     ImU32 col = ImGui::GetColorU32(ImVec4(1.0f, 0.85f, 0.2f, 0.5f + 0.5f * pulse));
